+++ src/components/WebcamButton.jsx
@@ -7,12 +7,11 @@ import { MdCallEnd } from "react-icons/md";
 import { BsCameraVideoFill, BsCameraVideoOffFill } from "react-icons/bs";
 import { AiTwotoneAudio, AiOutlineAudioMuted } from "react-icons/ai";
 
-const WebcamButton = () => {
+const WebcamButton = ({ isQuestionStarted }) => {
   const { isMuted, isWebcamOpen } = useStore();
 
   const showModalSmall = () => {
-    useStore.setState({ modalSize: "SMALL" });
-    useStore.setState({ showModal: true });
+    useStore.setState({ modalSize: "SMALL", showModal: true });
   };
 
   return (
@@ -28,6 +27,7 @@ const WebcamButton = () => {
         {isMuted ? <AiOutlineAudioMuted /> : <AiTwotoneAudio />}
       </Icon>
       <Icon
+        disabled={isQuestionStarted}
         style={
           !isWebcamOpen
             ? { backgroundColor: "#ea4335" }
@@ -37,7 +37,10 @@ const WebcamButton = () => {
       >
         {isWebcamOpen ? <BsCameraVideoFill /> : <BsCameraVideoOffFill />}
       </Icon>
-      <Icon onClick={useStore((state) => state.toggleIsMirrored)}>
+      <Icon
+        disabled={isQuestionStarted}
+        onClick={useStore((state) => state.toggleIsMirrored)}
+      >
         <FaExchangeAlt />
       </Icon>
       <Icon className="end" onClick={showModalSmall}>
@@ -78,10 +81,10 @@ const Icon = styled.button`
   font-size: 20px;
   border-radius: 50%;
   color: white;
-  cursor: pointer;
-  opacity: 0.9;
+  cursor: ${(props) => (props.disabled ? "" : "pointer")};
+  opacity: ${(props) => (props.disabled ? 0.5 : 0.9)};
 
-  :hover {
+  :${(props) => (props.disabled ? "" : "hover")} {
     opacity: 1;
     transform: scale(1.03);
   }
